% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_content}
\alias{get_content}
\title{Get WordPress Page or Post Content}
\usage{
get_content(id = NULL, wp_site, wp_header, content_type = "posts")
}
\arguments{
\item{id}{Numeric or \code{NULL}. The ID of the specific page or post to retrieve. If \code{NULL}, retrieves all content of the specified \code{content_type}.}

\item{wp_site}{Character. The base URL of the WordPress site (e.g., "https://example.com").}

\item{wp_header}{List. A named list containing the authorization headers required for API requests.}

\item{content_type}{Character. The type of content to retrieve, either "posts" (default) or "pages".}
}
\value{
A list containing the details of the retrieved content or an error message if the request fails.
}
\description{
Retrieves a page or post from a WordPress site using the REST API.
}
\details{
This function interacts with the WordPress REST API to retrieve content. By default, it fetches "posts", but setting \code{content_type = "pages"} will fetch pages instead. If \code{id} is provided, it retrieves a specific page or post by its ID. Otherwise, it retrieves a list of all posts or pages.
}
\examples{
\dontrun{
# Fetch all posts
get_content(wp_site = "https://example.com", wp_header = add_headers(Authorization = paste("Basic", wp_token)))

# Fetch a specific page by ID
get_content(id = 5, wp_site = "https://example.com", wp_header = add_headers(Authorization = paste("Basic", wp_token)), content_type = "pages")
}

}
