% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_content_pagination}
\alias{get_content_pagination}
\title{Fetch WordPress Posts with Pagination}
\usage{
get_content_pagination(wp_site, wp_header, content_type = "posts")
}
\arguments{
\item{wp_site}{Character. The base URL of the WordPress site (e.g., "https://example.com").}

\item{wp_header}{List. A named list created using \code{add_headers(Authorization = paste("Basic", wp_token))} for authentication, where \code{wp_token} is the base64-encoded username and password for the WordPress site.}

\item{content_type}{Character. The type of content to delete, either "posts" (default) or "pages".}
}
\value{
A list containing the posts retrieved for the specified page, or an error message if the request fails.
}
\description{
Retrieves a list of posts from a WordPress site, with support for pagination.
}
\details{
This function allows users to fetch posts from a WordPress site with pagination. By specifying the \code{per_page} and \code{page} parameters, users can control how many posts to retrieve and from which page. If the \code{wp_header} is provided, the function will include authentication in the API request.
}
\examples{
\dontrun{
# Fetch the first page of posts with the default number per page
get_content_pagination(wp_site = "https://example.com", wp_header = wp_header, content_type = "pages")
}

}
