% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_media}
\alias{get_media}
\title{Fetch WordPress Media Items}
\usage{
get_media(wp_site, wp_header, per_page = 10, page = 1)
}
\arguments{
\item{wp_site}{Character. The base URL of the WordPress site (e.g., "https://example.com").}

\item{wp_header}{List. A named list created using \code{add_headers(Authorization = paste("Basic", wp_token))} for authentication, where \code{wp_token} is the base64-encoded username and password for the WordPress site.}

\item{per_page}{Numeric. The number of media items to retrieve per page (default is 10).}

\item{page}{Numeric. The page number to retrieve (default is 1).}
}
\value{
A list containing the media items retrieved for the specified page, or an error message if the request fails.
}
\description{
Retrieves a list of media items from the WordPress media library using the REST API.
}
\details{
This function allows users to fetch media items from the WordPress media library via the REST API. By specifying the \code{per_page} and \code{page} parameters, users can control how many media items to retrieve and from which page. If the \code{wp_header} is provided, the function will include authentication in the API request.
}
\examples{
\dontrun{
# Fetch the first page of media items with the default number per page
get_media(wp_site = "https://example.com", wp_header = wp_header)

# Fetch the second page with 5 media items per page
get_media(wp_site = "https://example.com", wp_header = wp_header, per_page = 5, page = 2)
}

}
