% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{wp_api_request}
\alias{wp_api_request}
\title{Make API Requests to WordPress REST API}
\usage{
wp_api_request(method, wp_site, endpoint, wp_header, data = NULL)
}
\arguments{
\item{method}{Character. The HTTP method to be used for the request (e.g., "GET", "POST", "PUT", "DELETE").}

\item{wp_site}{Character. The base URL of the WordPress site (e.g., "https://example.com").}

\item{endpoint}{Character. The specific endpoint to which the request will be made (e.g., "wp/v2/posts").}

\item{wp_header}{List. Authentication headers created using \code{add_headers(Authorization = paste("Basic", wp_token))}.}

\item{data}{Optional. A list containing data to be sent with the request (for "POST" or "PUT" requests). Default is NULL.}
}
\value{
A list containing the response from the API request or an error message if the request fails.
}
\description{
This function performs HTTP requests to the WordPress REST API, supporting various HTTP methods.
}
\details{
This function is a general-purpose utility for making API requests to the WordPress REST API. It abstracts the underlying \code{httr} functions, allowing users to easily interact with various endpoints.
}
\examples{
\dontrun{
wp_header <- httr::add_headers(Authorization = paste("Basic", wp_token))

# Make a GET request to fetch posts
response <- wp_api_request("GET", wp_site = "https://example.com", endpoint = "wp/v2/posts", wp_header = wp_header)

# Make a POST request to create a new post
new_post_data <- list(title = "New Post", content = "This is the content of the new post.")
response <- wp_api_request("POST", wp_site = "https://example.com", endpoint = "wp/v2/posts", wp_header = wp_header, data = new_post_data)
}

}
